don't call s3.delete_objects with empty array

+++ app/jobs/projects_cleanup_job.rb
@@ -6,12 +6,10 @@ class ProjectsCleanupJob < ApplicationJob
   def perform()
     puts "cleanup ran at #{Time.now}"
 
-    Project.where("updated_at < :date", {date: Time.now - 48*3600}).find_each do |project|
+    Project.where("updated_at < :date", {date: Time.now - 168*3600}).find_each do |project|
       hashid = project.hashid
 
       puts "cleanup PURGE id=#{hashid}"
-
-      project.destroy
 
       s3 = MyS3Client.get
 
@@ -20,12 +18,18 @@ class ProjectsCleanupJob < ApplicationJob
         prefix: hashid
       })
 
+      if resp.contents.length == 0
+        next
+      end
+
       s3.delete_objects({
         bucket: MyS3Client.bucketName,
         delete: {
           objects: resp.contents.map { |it| { key: it.key } }
         }
       })
+
+      project.destroy
     end
   end
 end
